Extract translation even logic to a hook
